+++ src/main/java/com/anuradha/fcmpushnotification/model/User.java
@@ -8,7 +8,7 @@ import java.util.List;
 public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int user_id;
+    private long user_id;
     private String name;
     private String deviceToken;
     @ManyToMany(mappedBy = "users")
@@ -33,11 +33,11 @@ public class User {
                 '}';
     }
 
-    public int getUser_id() {
+    public long getUser_id() {
         return user_id;
     }
 
-    public void setUser_id(int user_id) {
+    public void setUser_id(long user_id) {
         this.user_id = user_id;
     }
